+++ android/src/main/java/com/synerise/synerise_flutter_sdk/SyneriseMethodChannel.java
@@ -1,5 +1,8 @@
 package com.synerise.synerise_flutter_sdk;
 
+import android.os.Handler;
+import android.util.Log;
+
 import io.flutter.embedding.engine.FlutterEngine;
 import io.flutter.plugin.common.MethodChannel;
 
@@ -8,5 +11,16 @@ public class SyneriseMethodChannel {
 
     public static void configureChannel(FlutterEngine flutterEngine) {
         methodChannel = new MethodChannel(flutterEngine.getDartExecutor().getBinaryMessenger(), "synerise_dart_channel");
+    }
+
+    public static void delayedMethodCall(MethodChannel methodChannel, String methodName, Object arguments, long delayMs) {
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                if (methodChannel != null) {
+                    methodChannel.invokeMethod(methodName, arguments);
+                }
+            }
+        }, delayMs);
     }
 }